feat: Add tilemaker to PATH.
By creating a symlink from `/tilemaker` to `/usr/bin/tilemaker`, one can execute the `tilemaker` command e.g. from within a shell.

This makes it possible to use the built tilemaker binary as if it was "installed". If one does a Docker multistep build that uses the tilemaker Docker image, one can then simly use `RUN tilemaker ...` and does not need to know where the binary is actually stored.