shard the node store to reduce memory usage
It looks like PR #499 adopts 36 bits as the max size of an OSM ID.

The NodeStore currently uses a full 64 bits for these IDs. This PR
changes it to shard the nodes across 16 collections (4 bits) and
then store only the last 32 bits in the collection itself.

This reduces memory usage for the NodeStore by 25%, without much
impact on runtime.

The CompactNodeStore is still much better, as it has no overhead and
constant time lookups -- but I'm often lazy and not using a renumbered
PBF file.