sigh.
I helpfully decided it'd be good if the copy constructor for the
intValues case initialized shortValues to all 0s to aid in debugging.

This forgets, of course, that shortValues is a union with intValues,
and so what we're really doing is stomping on the pointers of the
vector we just allocated and populated. Oops.