tile_data: ignore objects w/invalid coordinates
Some objects in Antarctica have latps that are out of range, causing
them to translate to a tile index that is invalid -- eg 16388, when
the range for z14 is from 0..16383.

I think the old code would silently accept such invalid coordinates,
insert them into a map, and nothing would read them and they'd be
dropped from the map.

This change also drops them from the map, but also writes a warning
to stderr.