update map for z9
Oops, I was thinking z9 was the base zoom, so it didn't make
sense to track tiles there--but z9 is the base zoom only for this
bitset, not for the actual map.

New timings:

real	1m51.573s
user	16m31.069s
sys	0m9.744s

This suggests that maybe it's worthwhile to track a secondary hierarchy
from z10..z13, too.