fix the bug, extend to z10..z13
Unfortunately, the bug turned out to be a critical part of the speed. :)

With it fixed, the new runtime is 11m5s. This is still a ~50% decrease,
but not what I was hoping for.

Some additional analysis shows that the frequency of a tile being entirely
excluded or covered is rarer than I initially thought -- it looks like
it's maybe 15-25% for Hudson Bay and James Bay.

I suspect a better savings would be to also keep the unsimplified, uncombined
clip from lower zooms around, and have higher zooms clip against them rather
the raw polygon.

I'll see if I can prototype that.