clipping monster polygons, take 2
This replaces https://github.com/systemed/tilemaker/pull/606

Rather than dealing with the special case of fully covered or fully
excluded, we keep a cache of recently clipped geometries.

A higher zoom will look in the cache for a clipped geometry of
the object from a lower zoom and re-use it if possible, falling
back to the whole geometry only if no cache entry is found.

Hudson Bay, master: 22m27s
this PR: 1m28s

This relies on the assumption that clipping a clipped polygon gives
the same result as clipping the original polygon. That seems
reasonable to me, but if you know of edge cases to consider,
let me know.