Clipping monster polygons 2 (#607)
* clipping monster polygons, take 2

This replaces https://github.com/systemed/tilemaker/pull/606

Rather than dealing with the special case of fully covered or fully
excluded, we keep a cache of recently clipped geometries.

A higher zoom will look in the cache for a clipped geometry of
the object from a lower zoom and re-use it if possible, falling
back to the whole geometry only if no cache entry is found.

Hudson Bay, master: 22m27s
this PR: 1m28s

This relies on the assumption that clipping a clipped polygon gives
the same result as clipping the original polygon. That seems
reasonable to me, but if you know of edge cases to consider,
let me know.

* more robustness against invalid Antarctica nodes

This is another fix in the vein of https://github.com/systemed/tilemaker/pull/595/commits/066ca0a91ea2981f55e63cd616af07a5dfff3125

Testing Antarctica with this PR: 9m5s

The generated mbtiles doesn't seem to be convertible to a pmtiles file
-- `pmtiles convert` fails with:

```
main.go:162: Failed to convert /home/cldellow/src/basemap/tiles.mbtiles, Missing row
```

I've never generated Antarctica with the previous code, so I'm not sure
if this is specific to this PR or not.

My intuition is this is not specific to this PR. Wild guess is that
we're creating tiles with no features, due to https://github.com/systemed/tilemaker/blob/d470dc94fea6ae74e948c1a858c6e1228c9f4bf9/src/tile_worker.cpp#L210, and pmtiles doesn't like that.