wip: use protozero for osm pbf reading
This is the minimal set of changes to use protozero, and is quite ugly
in places.

It also still devolves to using std::string for tags, so we're still
paying a cost for allocating long strings on the heap. I hope to look
at that next.

Time for the PBF read phase:

GB
  planet-on-32gb: 1m5s, 1GB RAM
  this branch: 55.7s, 0.74G RAM

NA
  planet-on-32gb: 10m49s, 3.01gb
  this branch: 10m10s, 2.68gb

So 6-15% speedup in reading, and perhaps a fixed 250MB reduction in
time.

The NA case is heavily affected by the Hudson Bay relation being a
straggler. Somehow tweaking things so that geometry could be read first
would ensure that the gains from protozero aren't lost to the whims
of the order of blocks in the PBF.