use protozero::data_view for tags
This didn't provide much impact on runtime or memory use, unfortunately.

Also use a batchSize of 1 for relations. Unlike nodes and ways, we're
not building a store, and so don't need long runs of continuous blocks.
This also hopefully decreases the odds of two slow blocks being assigned
to the same worker.