openmaptiles: don't pass false to LayerAsCentroid
Ugh, I guess this could be considered a breaking change after all.

If a user previously passed `LayerAsCentroid("layername", false)`,
we'd ignore the false.

With this change, the false causes us to fail, because we expect a
string.

I'd normally just say this was undefined behaviour, and those users
deserve to be broken...but it's tricky, since this is a script from the
official tilemaker repo.