compile streamvbyte as C, not C++
In #590, I compiled these as C++ files.

This was wrong, as they're meant to be compiled as C files. Depending on
your compiler, you would hit errors to do with `restrict` or
`__restrict__` references.

This reverts the streamvbyte code to be a more vanilla copy of upstream,
and updates the `Makefile` and `CMakeLists.txt` files to build them as C.